NL2SQL_System/tests/test_database.py
"""Tests for database connection."""
import pytest
from database.connection import MySQLConnection


def test_connection_initialization():
    """Test that connection manager initializes properly."""
    db = MySQLConnection()
    assert db.connection_params['host'] is not None
    assert db.connection_params['database'] is not None


# Note: Actual connection tests require a valid MySQL database
# These would be integration tests that run against a test database
def test_placeholder():
    """Placeholder test."""
    assert True
