NL2SQL_System/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

langchain>=0.2.0
langchain-openai>=0.1.0

pydantic>=2.0.0
pydantic-settings>=2.0.0

python-dotenv>=1.0.0
loguru>=0.7.0

cryptography>=42.0.0
pyjwt>=2.8.0

pymysql>=1.1.0
sqlglot>=28.0.0
sqlparse>=0.5.0

requests>=2.31.0
pandas>=2.0.0

# Privacy & PII Detection
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.0

streamlit>=1.28.0
plotly>=5.18.0
redis>=5.0.0